remove get-layout file
