+++ packages/hash/frontend/src/shared/layout.tsx
@@ -1,8 +1,26 @@
 import { NextPage } from "next";
 import { ReactElement, ReactNode } from "react";
+import { LayoutWithHeader } from "./layout/layout-with-header";
+import { LayoutWithSidebar } from "./layout/layout-with-sidebar";
+import { SidebarContextProvider } from "./layout/layout-with-sidebar/sidebar-context";
+import { PlainLayout } from "./layout/plain-layout";
 
 export type NextPageWithLayout<T = {}> = NextPage<T> & {
   getLayout?: (page: ReactElement) => ReactNode;
 };
 
-export * from "./layout/get-layout";
+export const getPlainLayout = (page: ReactElement) => {
+  return <PlainLayout>{page}</PlainLayout>;
+};
+
+export const getLayoutWithHeader = (page: ReactElement) => {
+  return <LayoutWithHeader>{page}</LayoutWithHeader>;
+};
+
+export const getLayoutWithSidebar = (page: ReactElement) => {
+  return (
+    <SidebarContextProvider>
+      <LayoutWithSidebar>{page}</LayoutWithSidebar>
+    </SidebarContextProvider>
+  );
+};
